[update] rm file & add it into cli
