+++ src/blurgenerator/cli.py
@@ -8,7 +8,6 @@ import cv2
 import numpy as np
 
 from blurgenerator import motion_blur, lens_blur, gaussian_blur
-from blurgenerator.depth_mapping import blur_with_depth_layers
 
 def main():
 
@@ -32,48 +31,79 @@ def main():
 
     args = parser.parse_args()
 
-    if args.input:
-        img_path = Path(args.input)
-        if img_path.is_file():
-            if img_path.suffix in ['.jpg', '.jpeg', '.png']:
-
-                img = cv2.imread(img_path.absolute().as_posix())
-                img = img / 255.
-
-                if args.type not in ['motion', 'lens', 'gaussian']:
-                    print('----- No type has been selected. Please specific `motion`, `lens`, or `gaussian`.')
-                else:
-                    if args.type == 'motion':
-                        def blur_job(img, size=args.motion_blur_size):
-                            return motion_blur(img, size=size, angle=args.motion_blur_angle)
-
-                    elif args.type == 'lens':
-                        def blur_job(img, radius=args.lens_radius):
-                            return lens_blur(img, radius=radius, components=args.lens_components, exposure_gamma=args.lens_exposure_gamma)
-
-                    elif args.type == 'gaussian':
-                        def blur_job(img, kernel=args.gaussian_kernel):
-                            return gaussian_blur(img, kernel)
-
-                    if args.type in ['motion', 'lens', 'gaussian']:
-                        depth_map_path = args.input_depth_map
-                        if depth_map_path is None:
-                            result = blur_job(img)
-                        else:
-                            result = np.zeros_like(img)
-                            # need check path exists
-                            depth_map = cv2.imread(depth_map_path)
-                            mask_blur_amounts = blur_with_depth_layers(depth_map)
-                            for mask, blur_amount in mask_blur_amounts:
-                                slice = blur_job(img, blur_amount)
-                                layer = cv2.bitwise_and(slice, slice, mask = mask[:,:,0])
-                                result = cv2.add(result, layer, dtype=0)
-                        cv2.imwrite(args.output, result)
-
-            else:
-                print('----- Only support common types of image `.jpg` and `.png`.')
-
-        else:
-            print('----- File not exists!')
-    else:
+    if not args.input:
         print('----- Please specific image for input.')
+        return
+    img_path = Path(args.input)
+
+    if not img_path.is_file():
+        print('----- `img_path` is not a file!')
+        return
+    if img_path.suffix not in ['.jpg', '.jpeg', '.png']:
+        print('----- Only support common types of image `.jpg` and `.png`.')
+        return
+
+    img = cv2.imread(img_path.absolute().as_posix())
+    img = img / 255.
+
+    if args.type not in ['motion', 'lens', 'gaussian']:
+        print('----- No type has been selected. Please specific `motion`, `lens`, or `gaussian`.')
+        return
+
+    if args.type == 'motion':
+        def blur_job(img, size=args.motion_blur_size):
+            return motion_blur(img, size=size, angle=args.motion_blur_angle)
+
+    if args.type == 'lens':
+        def blur_job(img, radius=args.lens_radius):
+            return lens_blur(img, radius=radius, components=args.lens_components, exposure_gamma=args.lens_exposure_gamma)
+
+    if args.type == 'gaussian':
+        def blur_job(img, kernel=args.gaussian_kernel):
+            return gaussian_blur(img, kernel)
+
+    depth_map_path = args.input_depth_map
+    if depth_map_path is None:
+        result = blur_job(img)
+        cv2.imwrite(args.output, result*255)
+        return
+
+    depth_map_path = Path(depth_map_path)
+    if not img_path.is_file():
+        print('----- `input_depth_map` is not a file!')
+        return
+    if depth_map_path.suffix not in ['.jpg', '.jpeg', '.png']:
+        print('----- Only support common types of image `.jpg` and `.png`.')
+        return
+
+    depth_map = cv2.imread(depth_map_path.absolute().as_posix())
+    def map_range(value, inMin, inMax, outMin, outMax):
+        return outMin + (((value - inMin) / (inMax - inMin)) * (outMax - outMin))
+
+    def blur_with_depth(img, depth, num_layers=10, min_blur=1, max_blur=100):
+        min_depth = np.min(np.unique(depth))
+        max_depth = np.max(np.unique(depth))
+        step = (max_depth - min_depth) // num_layers
+        layers = np.array(range(min_depth, max_depth, step))
+        out = np.zeros(img.shape)
+
+        for value in layers:
+            dm = cv2.cvtColor(depth, cv2.COLOR_BGR2GRAY)
+            m = np.zeros(dm.shape)
+            m[dm > value] = 255
+            m[dm > (value + step)] = 0
+            l_mask = depth.copy()
+            l_mask[:,:,0] = m[:,:]
+            l_mask[:,:,1] = m[:,:]
+            l_mask[:,:,2] = m[:,:]
+            blur_amount = int(map_range(value, 0, 255, min_blur, max_blur))
+            slice = blur_job(img, blur_amount)
+            _, mask = cv2.threshold(l_mask, 100, 255, cv2.THRESH_BINARY)
+            layer = cv2.bitwise_and(slice, slice, mask = mask[:,:,0])
+            out = cv2.add(out, layer, dtype=0)
+        return out
+
+    result = blur_with_depth(img, depth_map, num_layers=5, min_blur=1, max_blur=50)
+    cv2.imwrite(args.output, result)
+
+    return
